Major Update - UX and API
Added a lot more error information to the page, new component Timer used to give information with timing functions, lot of progress on the administration and teacher part of the system. Modified some entities and added some to better reflect expected output and made some progress on the Subject page.